Привязка к языку в схеме запуска для Debug попала в репозиторий, вернул System Language. Исправил другие замечания по итогам ревью
