+++ Tracker/TrackerTests/TrackerTests.swift
@@ -14,11 +14,11 @@ final class TrackerTests: XCTestCase {
     func testStartViewController() throws {
         let startVC = StartViewController()
         // .light mode
-        assertSnapshots(of: startVC, as: [.image(on: .iPhone13ProMax, traits: .init(userInterfaceStyle: .light))])
-        assertSnapshots(of: startVC, as: [.image(on: .iPhone13, traits: .init(userInterfaceStyle: .light))])
+        assertSnapshots(of: startVC, as: [.image(on: .iPhone13ProMax, traits: UITraitCollection(userInterfaceStyle: .light))])
+        assertSnapshots(of: startVC, as: [.image(on: .iPhone13, traits: UITraitCollection(userInterfaceStyle: .light))])
 
         // .dark mode
-        assertSnapshots(of: startVC, as: [.image(on: .iPhone13ProMax, traits: .init(userInterfaceStyle: .dark))])
-        assertSnapshots(of: startVC, as: [.image(on: .iPhone13, traits: .init(userInterfaceStyle: .dark))])
+        assertSnapshots(of: startVC, as: [.image(on: .iPhone13ProMax, traits: UITraitCollection(userInterfaceStyle: .dark))])
+        assertSnapshots(of: startVC, as: [.image(on: .iPhone13, traits: UITraitCollection(userInterfaceStyle: .dark))])
     }
 }
